travel_backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.4.3
PyMySQL==1.0.2
python-dotenv==0.20.0
